@@ -1,4 +1,4 @@
 export * from './config.config';
 export * from './database.config';
 export * from './ogma.config';
-export * from './passport.config';
+export * from './redis.config';
